organise Pool types by context
