+++ src/contexts/Pools/PoolMemberships/defaults.ts
@@ -1,7 +1,7 @@
 // Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { PoolMembership, PoolMembershipsContextState } from '../types';
+import type { PoolMembership, PoolMembershipsContextState } from './types';
 
 export const poolMembership: PoolMembership | null = null;
