+++ src/library/Hooks/usePoolFilters/index.tsx
@@ -2,10 +2,10 @@
 
 import { useTranslation } from 'react-i18next';
 import { useBondedPools } from 'contexts/Pools/BondedPools';
-import type { BondedPool } from 'contexts/Pools/types';
 import { useStaking } from 'contexts/Staking';
 import type { AnyFunction, AnyJson } from 'types';
 import type { AnyFilter } from 'library/Filter/types';
+import type { BondedPool } from 'contexts/Pools/BondedPools/types';
 
 export const usePoolFilters = () => {
   const { t } = useTranslation('library');
