+++ src/library/Pool/types.ts
@@ -1,7 +1,8 @@
 // Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { PoolAddresses, PoolRoles, PoolState } from 'contexts/Pools/types';
+import type { PoolRoles, PoolState } from 'contexts/Pools/ActivePools/types';
+import type { PoolAddresses } from 'contexts/Pools/BondedPools/types';
 import type { DisplayFor } from 'types';
 
 export interface PoolProps {
