+++ src/modals/ManagePool/Forms/SetClaimPermission/index.tsx
@@ -12,7 +12,6 @@ import { useTranslation } from 'react-i18next';
 import { useApi } from 'contexts/Api';
 import { useActivePools } from 'contexts/Pools/ActivePools';
 import { usePoolMemberships } from 'contexts/Pools/PoolMemberships';
-import type { ClaimPermission } from 'contexts/Pools/types';
 import { ClaimPermissionInput } from 'library/Form/ClaimPermissionInput';
 import { Warning } from 'library/Form/Warning';
 import { useSignerWarnings } from 'library/Hooks/useSignerWarnings';
@@ -20,6 +19,7 @@ import { useSubmitExtrinsic } from 'library/Hooks/useSubmitExtrinsic';
 import { SubmitTx } from 'library/SubmitTx';
 import { useOverlay } from '@polkadot-cloud/react/hooks';
 import { useActiveAccounts } from 'contexts/ActiveAccounts';
+import type { ClaimPermission } from 'contexts/Pools/PoolMemberships/types';
 
 export const SetClaimPermission = ({ setSection, section }: any) => {
   const { t } = useTranslation('modals');
