+++ src/pages/Overview/NetworkSats/Announcements.tsx
@@ -14,12 +14,12 @@ import { motion } from 'framer-motion';
 import { useTranslation } from 'react-i18next';
 import { useBondedPools } from 'contexts/Pools/BondedPools';
 import { usePoolsConfig } from 'contexts/Pools/PoolsConfig';
-import type { BondedPool } from 'contexts/Pools/types';
 import { useStaking } from 'contexts/Staking';
 import { useUi } from 'contexts/UI';
 import { Announcement as AnnouncementLoader } from 'library/Loader/Announcement';
 import { useNetwork } from 'contexts/Network';
 import { Item } from './Wrappers';
+import type { BondedPool } from 'contexts/Pools/BondedPools/types';
 
 export const Announcements = () => {
   const { t } = useTranslation('pages');
